# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CodeInput {
  /** JavaScript source code to execute */
  code: string;
}

export interface CodeResult {
  /** Captured console output */
  output: string;
  /**
     * Error message if execution failed
     * @nullable
     */
  error: string | null;
  /** Execution time in milliseconds */
  executionTime: number;
}

export type ProblemDifficulty = typeof ProblemDifficulty[keyof typeof ProblemDifficulty];


export const ProblemDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface Problem {
  id: string;
  title: string;
  difficulty: ProblemDifficulty;
  category: string;
  tags: string[];
}

export type ProblemDetailDifficulty = typeof ProblemDetailDifficulty[keyof typeof ProblemDetailDifficulty];


export const ProblemDetailDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface TestCase {
  input: string;
  expected: string;
}

export interface ProblemDetail {
  id: string;
  title: string;
  difficulty: ProblemDetailDifficulty;
  category: string;
  tags: string[];
  description: string;
  starterCode: string;
  testCases: TestCase[];
}

export type ProblemStatsByDifficulty = {[key: string]: number};

export type ProblemStatsByCategory = {[key: string]: number};

export interface ProblemStats {
  total: number;
  byDifficulty: ProblemStatsByDifficulty;
  byCategory: ProblemStatsByCategory;
}

export interface Snippet {
  id: number;
  title: string;
  code: string;
  createdAt: string;
}

export interface SnippetInput {
  title: string;
  code: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProblemsParams = {
category?: string;
difficulty?: string;
};