# lib/api-zod/src/generated/types/codeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeInput {
  /** JavaScript source code to execute */
  code: string;
}