# lib/api-zod/src/generated/types/codeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeResult {
  /** Captured console output */
  output: string;
  /**
     * Error message if execution failed
     * @nullable
     */
  error: string | null;
  /** Execution time in milliseconds */
  executionTime: number;
}