# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemDifficulty } from './problemDifficulty';

export interface Problem {
  id: string;
  title: string;
  difficulty: ProblemDifficulty;
  category: string;
  tags: string[];
}