# lib/api-zod/src/generated/types/problemDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemDetailDifficulty } from './problemDetailDifficulty';
import type { TestCase } from './testCase';

export interface ProblemDetail {
  id: string;
  title: string;
  difficulty: ProblemDetailDifficulty;
  category: string;
  tags: string[];
  description: string;
  starterCode: string;
  testCases: TestCase[];
}