# lib/api-zod/src/generated/types/problemStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemStatsByCategory } from './problemStatsByCategory';
import type { ProblemStatsByDifficulty } from './problemStatsByDifficulty';

export interface ProblemStats {
  total: number;
  byDifficulty: ProblemStatsByDifficulty;
  byCategory: ProblemStatsByCategory;
}