# lib/api-zod/src/generated/types/snippet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JS Playground API
 * OpenAPI spec version: 0.1.0
 */

export interface Snippet {
  id: number;
  title: string;
  code: string;
  createdAt: string;
}